Rename skill paths and fix crest positions
