+++ src/App.tsx
@@ -2,20 +2,20 @@ import { useState } from 'react';
 import LandingPage from './LandingPage';
 import WarriorPage from './WarriorPage';
 import SkaldPage from './SkaldPage';
-import BerserkerPage from './BerserkerPage';
+import SmithPage from './SmithPage';
 
-export type Page = 'landing' | 'warrior' | 'skald' | 'berserker';
+export type Page = 'landing' | 'smith' | 'skald' | 'warrior';
 
 const App = () => {
   const [page, setPage] = useState<Page>('landing');
 
   switch (page) {
-    case 'warrior':
-      return <WarriorPage onBack={() => setPage('landing')} />;
+    case 'smith':
+      return <SmithPage onBack={() => setPage('landing')} />;
     case 'skald':
       return <SkaldPage onBack={() => setPage('landing')} />;
-    case 'berserker':
-      return <BerserkerPage onBack={() => setPage('landing')} />;
+    case 'warrior':
+      return <WarriorPage onBack={() => setPage('landing')} />;
     default:
       return <LandingPage onSelect={setPage} />;
   }
